refactor(request): rename request to network class
